refactor: split go in multiple files
