+++ docker/cs-web-server/src/server/webrtc.go
@@ -1,0 +1,535 @@
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"io"
+	"math/rand"
+	"net/http"
+	"os"
+	"strconv"
+	"sync"
+	"time"
+
+	"github.com/gorilla/websocket"
+	"github.com/pion/logging"
+	"github.com/pion/rtcp"
+	"github.com/pion/rtp"
+	"github.com/pion/webrtc/v4"
+	goxash3d_fwgs "github.com/yohimik/goxash3d-fwgs/pkg"
+)
+
+var (
+	upgrader = websocket.Upgrader{
+		CheckOrigin: func(r *http.Request) bool { return true },
+	}
+
+	api *webrtc.API
+
+	// lock for peerConnections and trackLocals
+	listLock        sync.RWMutex
+	peerConnections []*peerConnectionState
+	trackLocals     map[string]*webrtc.TrackLocalStaticRTP
+
+	log = logging.NewDefaultLoggerFactory().NewLogger("sfu-ws")
+)
+
+const DefaultSignalsCount = 5
+const messageSize = 1024 * 8
+
+const DefaultPongWaitSeconds = 60
+
+func parsePongWait() time.Duration {
+	v := os.Getenv("PONG_WAIT_SECONDS")
+	s, err := strconv.Atoi(v)
+	if err != nil {
+		return DefaultPongWaitSeconds
+	}
+	if s < 1 {
+		return DefaultPongWaitSeconds
+	}
+	return time.Duration(s)
+}
+
+const DefaultWriteWaitSeconds = 10
+
+func parseWriteWait() time.Duration {
+	v := os.Getenv("WRITE_WAIT_SECONDS")
+	s, err := strconv.Atoi(v)
+	if err != nil {
+		return DefaultWriteWaitSeconds
+	}
+	if s < 1 {
+		return DefaultWriteWaitSeconds
+	}
+	return time.Duration(s)
+}
+
+// Websocket keepalive tuning. pongWait is how long we'll wait for a pong
+// (or any other message) before considering the connection dead. pingPeriod
+// must be less than pongWait so pings go out with time to spare.
+var (
+	pongWait   = parsePongWait() * time.Second
+	pingPeriod = (pongWait * 9) / 10
+	writeWait  = parseWriteWait() * time.Second
+)
+
+type websocketMessage struct {
+	Event string          `json:"event"`
+	Data  json.RawMessage `json:"data"`
+}
+
+type peerConnectionState struct {
+	peerConnection *webrtc.PeerConnection
+	websocket      *threadSafeWriter
+	signalsCount   int
+}
+
+// Helper to make Gorilla Websockets threadsafe.
+type threadSafeWriter struct {
+	*websocket.Conn
+	sync.Mutex
+}
+
+func (t *threadSafeWriter) WriteJSON(event string, v interface{}) error {
+	t.Lock()
+	defer t.Unlock()
+
+	t.Conn.SetWriteDeadline(time.Now().Add(writeWait)) // nolint
+
+	return t.Conn.WriteJSON(struct {
+		Event string `json:"event"`
+		Data  any    `json:"data"`
+	}{event, v})
+}
+
+// Ping sends a websocket control-frame ping. Pairs with the pong handler
+// registered in websocketHandler to detect and clean up dead connections.
+func (t *threadSafeWriter) Ping() error {
+	t.Lock()
+	defer t.Unlock()
+
+	return t.Conn.WriteControl(websocket.PingMessage, nil, time.Now().Add(writeWait))
+}
+
+// Add to list of tracks and fire renegotation for all PeerConnections.
+func addTrack(t *webrtc.TrackRemote) *webrtc.TrackLocalStaticRTP { // nolint
+	listLock.Lock()
+	defer func() {
+		listLock.Unlock()
+		signalPeerConnections()
+	}()
+
+	// Create a new TrackLocal with the same codec as our incoming
+	trackLocal, err := webrtc.NewTrackLocalStaticRTP(t.Codec().RTPCodecCapability, t.ID(), t.StreamID())
+	if err != nil {
+		panic(err)
+	}
+
+	trackLocals[t.ID()] = trackLocal
+
+	for _, con := range peerConnections {
+		con.signalsCount = DefaultSignalsCount
+	}
+
+	return trackLocal
+}
+
+// Remove from list of tracks and fire renegotation for all PeerConnections.
+func removeTrack(t *webrtc.TrackLocalStaticRTP) {
+	listLock.Lock()
+	defer func() {
+		listLock.Unlock()
+		signalPeerConnections()
+	}()
+
+	for _, con := range peerConnections {
+		con.signalsCount = DefaultSignalsCount
+	}
+
+	delete(trackLocals, t.ID())
+}
+
+// signalPeerConnections updates each PeerConnection so that it is getting all the expected media tracks.
+func signalPeerConnections() { // nolint
+	listLock.Lock()
+	defer func() {
+		listLock.Unlock()
+		dispatchKeyFrame()
+	}()
+
+	attemptSync := func() (tryAgain bool) {
+		for i := range peerConnections {
+			if peerConnections[i].signalsCount <= 0 {
+				continue
+			}
+
+			if peerConnections[i].peerConnection.ConnectionState() == webrtc.PeerConnectionStateClosed {
+				peerConnections = append(peerConnections[:i], peerConnections[i+1:]...)
+
+				return true // We modified the slice, start from the beginning
+			}
+
+			// map of sender we already are seanding, so we don't double send
+			existingSenders := map[string]bool{}
+
+			for _, sender := range peerConnections[i].peerConnection.GetSenders() {
+				if sender.Track() == nil {
+					continue
+				}
+
+				existingSenders[sender.Track().ID()] = true
+
+				// If we have a RTPSender that doesn't map to a existing track remove and signal
+				if _, ok := trackLocals[sender.Track().ID()]; !ok {
+					if err := peerConnections[i].peerConnection.RemoveTrack(sender); err != nil {
+						return true
+					}
+				}
+			}
+
+			// Don't receive videos we are sending, make sure we don't have loopback
+			for _, receiver := range peerConnections[i].peerConnection.GetReceivers() {
+				if receiver.Track() == nil {
+					continue
+				}
+
+				existingSenders[receiver.Track().ID()] = true
+			}
+
+			// Add all track we aren't sending yet to the PeerConnection
+			for trackID := range trackLocals {
+				if _, ok := existingSenders[trackID]; !ok {
+					if _, err := peerConnections[i].peerConnection.AddTrack(trackLocals[trackID]); err != nil {
+						return true
+					}
+				}
+			}
+
+			offer, err := peerConnections[i].peerConnection.CreateOffer(nil)
+			if err != nil {
+				return true
+			}
+
+			if err = peerConnections[i].peerConnection.SetLocalDescription(offer); err != nil {
+				return true
+			}
+
+			if err = peerConnections[i].websocket.WriteJSON("offer", offer); err != nil {
+				return true
+			}
+		}
+
+		return tryAgain
+	}
+
+	for syncAttempt := 0; ; syncAttempt++ {
+		if syncAttempt == 25 {
+			// Release the lock and attempt a sync in 3 seconds. We might be blocking a RemoveTrack or AddTrack
+			go func() {
+				time.Sleep(time.Second * 3)
+				signalPeerConnections()
+			}()
+
+			return
+		}
+
+		if !attemptSync() {
+			break
+		}
+	}
+}
+
+// dispatchKeyFrame sends a keyframe to all PeerConnections, used everytime a new user joins the call.
+func dispatchKeyFrame() {
+	listLock.Lock()
+	defer listLock.Unlock()
+
+	for i := range peerConnections {
+		for _, receiver := range peerConnections[i].peerConnection.GetReceivers() {
+			if receiver.Track() == nil {
+				continue
+			}
+
+			_ = peerConnections[i].peerConnection.WriteRTCP([]rtcp.Packet{
+				&rtcp.PictureLossIndication{
+					MediaSSRC: uint32(receiver.Track().SSRC()),
+				},
+			})
+		}
+	}
+}
+
+func ReadLoop(d io.Reader, ip [4]byte) {
+	for {
+		buffer := make([]byte, messageSize)
+		n, err := d.Read(buffer)
+		if err != nil {
+			fmt.Println("Datachannel closed; Exit the readloop:", err)
+
+			return
+		}
+		net.PushPacket(goxash3d_fwgs.Packet{
+			Addr: goxash3d_fwgs.Addr{
+				IP:   ip,
+				Port: 1000,
+			},
+			Data: buffer[:n],
+		})
+	}
+}
+
+// Handle incoming websockets.
+func websocketHandler(w http.ResponseWriter, r *http.Request) { // nolint
+	// Upgrade HTTP request to Websocket
+	unsafeConn, err := upgrader.Upgrade(w, r, nil)
+	if err != nil {
+		log.Errorf("Failed to upgrade HTTP to Websocket: ", err)
+
+		return
+	}
+
+	c := &threadSafeWriter{unsafeConn, sync.Mutex{}} // nolint
+
+	// When this frame returns close the Websocket
+	defer c.Close() //nolint
+
+	// --- Ping/pong keepalive setup ---
+	// Track liveness: any pong (or other) frame from the client resets the
+	// read deadline. If the client stops responding, reads will time out and
+	// the handler will exit, which triggers the deferred cleanup.
+	c.SetReadDeadline(time.Now().Add(pongWait)) // nolint
+	c.SetPongHandler(func(string) error {
+		c.SetReadDeadline(time.Now().Add(pongWait)) // nolint
+		return nil
+	})
+
+	pingDone := make(chan struct{})
+	defer close(pingDone)
+
+	go func() {
+		ticker := time.NewTicker(pingPeriod)
+		defer ticker.Stop()
+
+		for {
+			select {
+			case <-ticker.C:
+				if err := c.Ping(); err != nil {
+					log.Errorf("Failed to send ping: %v", err)
+					return
+				}
+			case <-pingDone:
+				return
+			}
+		}
+	}()
+	// --- end ping/pong keepalive setup ---
+
+	// Create new PeerConnection
+	peerConnection, err := api.NewPeerConnection(webrtc.Configuration{})
+	if err != nil {
+		log.Errorf("Failed to creates a PeerConnection: %v", err)
+
+		return
+	}
+
+	// When this frame returns close the PeerConnection
+	defer peerConnection.Close() //nolint
+
+	// Accept one audio track incoming
+	for _, typ := range []webrtc.RTPCodecType{webrtc.RTPCodecTypeAudio} {
+		if _, err := peerConnection.AddTransceiverFromKind(typ, webrtc.RTPTransceiverInit{
+			Direction: webrtc.RTPTransceiverDirectionRecvonly,
+		}); err != nil {
+			log.Errorf("Failed to add transceiver: %v", err)
+
+			return
+		}
+	}
+
+	f := false
+	var z uint16 = 0
+	if err != nil {
+		log.Errorf("Failed to creates a data channel: %v", err)
+
+		return
+	}
+	ip := [4]byte{}
+	for i := range ip {
+		ip[i] = byte(rand.Intn(256))
+	}
+	index, _ := pool.TryGet()
+	ip[0] = index
+	defer pool.TryPut(index)
+
+	writeChannel, err := peerConnection.CreateDataChannel("write", &webrtc.DataChannelInit{
+		Ordered:        &f,
+		MaxRetransmits: &z,
+	})
+	if err != nil {
+		log.Errorf("Failed to creates a data channel: %v", err)
+
+		return
+	}
+	var readChannel *webrtc.DataChannel
+	defer func() {
+		if readChannel != nil {
+			readChannel.Close()
+		}
+	}()
+	writeChannel.OnOpen(func() {
+		d, err := writeChannel.Detach()
+		if err != nil {
+			panic(err)
+		}
+		connections[index] = d
+
+		rc, err := peerConnection.CreateDataChannel("read", &webrtc.DataChannelInit{
+			Ordered:        &f,
+			MaxRetransmits: &z,
+		})
+		if err != nil {
+			log.Errorf("Failed to creates a data channel: %v", err)
+
+			return
+		}
+		readChannel = rc
+		readChannel.OnOpen(func() {
+			d, err := readChannel.Detach()
+			if err != nil {
+				panic(err)
+			}
+			go ReadLoop(d, ip)
+		})
+	})
+	defer writeChannel.Close()
+
+	// Trickle ICE. Emit server candidate to client
+	peerConnection.OnICECandidate(func(i *webrtc.ICECandidate) {
+		if i == nil {
+			return
+		}
+		// If you are serializing a candidate make sure to use ToJSON
+		// Using Marshal will result in errors around `sdpMid`
+
+		if writeErr := c.WriteJSON("candidate", i.ToJSON()); writeErr != nil {
+			log.Errorf("Failed to write JSON: %v", writeErr)
+		}
+	})
+
+	// If PeerConnection is closed remove it from global list
+	peerConnection.OnConnectionStateChange(func(p webrtc.PeerConnectionState) {
+		switch p {
+		case webrtc.PeerConnectionStateFailed:
+			if err := peerConnection.Close(); err != nil {
+				log.Errorf("Failed to close PeerConnection: %v", err)
+			}
+		case webrtc.PeerConnectionStateClosed:
+			signalPeerConnections()
+		default:
+		}
+	})
+
+	peerConnection.OnTrack(func(t *webrtc.TrackRemote, _ *webrtc.RTPReceiver) {
+		// Create a track to fan out our incoming video to all peers
+		trackLocal := addTrack(t)
+		defer removeTrack(trackLocal)
+
+		buf := make([]byte, 1500)
+		rtpPkt := &rtp.Packet{}
+
+		for {
+			i, _, err := t.Read(buf)
+			if err != nil {
+				return
+			}
+
+			if err = rtpPkt.Unmarshal(buf[:i]); err != nil {
+				log.Errorf("Failed to unmarshal incoming RTP packet: %v", err)
+
+				return
+			}
+
+			rtpPkt.Extension = false
+			rtpPkt.Extensions = nil
+
+			if err = trackLocal.WriteRTP(rtpPkt); err != nil {
+				return
+			}
+		}
+	})
+
+	// Add our new PeerConnection to global list
+	state := peerConnectionState{peerConnection, c, DefaultSignalsCount}
+	listLock.Lock()
+	peerConnections = append(peerConnections, &state)
+	listLock.Unlock()
+
+	// Signal for the new PeerConnection
+	signalPeerConnections()
+
+	message := &websocketMessage{}
+	for {
+		_, raw, err := c.ReadMessage()
+		if err != nil {
+			log.Errorf("Failed to read message: %v", err)
+
+			return
+		}
+
+		// Any inbound frame counts as liveness too, not just pongs.
+		c.SetReadDeadline(time.Now().Add(pongWait)) // nolint
+
+		if err := json.Unmarshal(raw, &message); err != nil {
+			log.Errorf("Failed to unmarshal json to message: %v", err)
+
+			return
+		}
+
+		switch message.Event {
+		case "candidate":
+			candidate := webrtc.ICECandidateInit{}
+			if err := json.Unmarshal(message.Data, &candidate); err != nil {
+				log.Errorf("Failed to unmarshal json to candidate: %v", err)
+
+				return
+			}
+
+			if err := peerConnection.AddICECandidate(candidate); err != nil {
+				log.Errorf("Failed to add ICE candidate: %v", err)
+
+				return
+			}
+		case "answer":
+			answer := webrtc.SessionDescription{}
+			if err := json.Unmarshal(message.Data, &answer); err != nil {
+				log.Errorf("Failed to unmarshal json to answer: %v", err)
+
+				return
+			}
+
+			if err := peerConnection.SetRemoteDescription(answer); err != nil {
+				log.Errorf("Failed to set remote description: %v", err)
+
+				return
+			}
+			listLock.Lock()
+			state.signalsCount -= 1
+			isNeedSignaling := state.signalsCount > 0
+			listLock.Unlock()
+			if isNeedSignaling {
+				signalPeerConnections()
+			}
+		case "ping":
+			// Application-level ping from the client (as opposed to the
+			// websocket control frame) - just echo a pong event back.
+			if err := c.WriteJSON("pong", nil); err != nil {
+				log.Errorf("Failed to write pong: %v", err)
+
+				return
+			}
+		default:
+			log.Errorf("unknown message: %+v", message)
+		}
+	}
+}
